move EnvironmentMixin to separate file
