+++ include/flox/search/command.hh
@@ -18,6 +18,7 @@
 #include "flox/pkgdb/write.hh"
 #include "flox/registry.hh"
 #include "flox/search/params.hh"
+#include "flox/resolver/mixins.hh"
 
 
 /* -------------------------------------------------------------------------- */
